+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Fix(es):
 - Add field serializers to `cpus`, `gpus`, `storages`, and `price_tiered` columns to prevent Pydantic serialization
   warnings.
 - Fix and extend test cases for field serializers and OVH vendor module.
+- Add Provisioned IOPS (io1, io2) storage types for AWS.
 
 New feature(s):